+++ sdk/examples/client/participation.rs
@@ -101,7 +101,7 @@ async fn participate(client: &Client, event_id: ParticipationEventId, node_url: 
     request_funds_from_faucet(&faucet_url, &address.to_bech32(client.get_bech32_hrp().await?)).await?;
 
     let outputs = vec![
-        BasicOutputBuilder::new_with_minimum_storage_deposit(rent_structure)?
+        BasicOutputBuilder::new_with_minimum_storage_deposit(rent_structure)
             .add_unlock_condition(AddressUnlockCondition::new(address))
             .finish_output(token_supply)?,
     ];