+++ sdk/src/wallet/account/builder.rs
@@ -6,8 +6,6 @@ use std::{
     sync::Arc,
 };
 
-#[cfg(feature = "events")]
-use tokio::sync::Mutex;
 use tokio::sync::RwLock;
 
 #[cfg(feature = "events")]
@@ -33,9 +31,9 @@ pub struct AccountBuilder {
     secret_manager: Arc<RwLock<SecretManager>>,
     accounts: Arc<RwLock<Vec<Account>>>,
     #[cfg(feature = "events")]
-    event_emitter: Arc<Mutex<EventEmitter>>,
+    event_emitter: Arc<tokio::sync::Mutex<EventEmitter>>,
     #[cfg(feature = "storage")]
-    storage_manager: Arc<Mutex<StorageManager>>,
+    storage_manager: Arc<tokio::sync::Mutex<StorageManager>>,
 }
 
 impl AccountBuilder {
@@ -45,8 +43,8 @@ impl AccountBuilder {
         client_options: Arc<RwLock<ClientOptions>>,
         coin_type: u32,
         secret_manager: Arc<RwLock<SecretManager>>,
-        #[cfg(feature = "events")] event_emitter: Arc<Mutex<EventEmitter>>,
-        #[cfg(feature = "storage")] storage_manager: Arc<Mutex<StorageManager>>,
+        #[cfg(feature = "events")] event_emitter: Arc<tokio::sync::Mutex<EventEmitter>>,
+        #[cfg(feature = "storage")] storage_manager: Arc<tokio::sync::Mutex<StorageManager>>,
     ) -> Self {
         Self {
             addresses: None,
@@ -65,20 +63,20 @@ impl AccountBuilder {
 
     /// Set the addresses, should only be used for accounts with an offline counterpart account from which the addresses
     /// were exported
-    pub fn with_addresses(mut self, addresses: Vec<AccountAddress>) -> Self {
-        self.addresses.replace(addresses);
+    pub fn with_addresses(mut self, addresses: impl Into<Option<Vec<AccountAddress>>>) -> Self {
+        self.addresses = addresses.into();
         self
     }
 
     /// Set the alias
-    pub fn with_alias(mut self, alias: String) -> Self {
-        self.alias.replace(alias);
+    pub fn with_alias(mut self, alias: impl Into<Option<String>>) -> Self {
+        self.alias = alias.into();
         self
     }
 
     /// Set the bech32 HRP
-    pub fn with_bech32_hrp(mut self, bech32_hrp: String) -> Self {
-        self.bech32_hrp.replace(bech32_hrp);
+    pub fn with_bech32_hrp(mut self, bech32_hrp: impl Into<Option<String>>) -> Self {
+        self.bech32_hrp = bech32_hrp.into();
         self
     }
 
@@ -214,6 +212,6 @@ pub(crate) async fn get_first_public_address(
     Ok(secret_manager
         .read()
         .await
-        .generate_addresses(coin_type, account_index, 0..1, false, None)
+        .generate_addresses(coin_type, account_index, 0..1, None)
         .await?[0])
 }