+++ sdk/src/wallet/account/operations/transaction/high_level/minting/mint_nfts.rs
@@ -151,7 +151,7 @@ impl Account {
 
             // NftId needs to be set to 0 for the creation
             let mut nft_builder =
-                NftOutputBuilder::new_with_minimum_storage_deposit(rent_structure.clone(), NftId::null())?
+                NftOutputBuilder::new_with_minimum_storage_deposit(rent_structure.clone(), NftId::null())
                     // Address which will own the nft
                     .add_unlock_condition(AddressUnlockCondition::new(address));