+++ sdk/CHANGELOG.md
@@ -29,6 +29,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `PostBlockPayload::payload_dto` renamed to `payload`;
 - `SendNativeTokens::addresses_native_tokens` renamed to `addresses_and_native_tokens`;
 - `SendNft::addresses_nft_ids` renamed to `addresses_and_nft_ids`;
+- `{AliasOutputBuilder, BasicOutputBuilder, FoundryOutputBuilder, NftOutputBuilder}::{new_with_amount, new_with_minimum_storage_deposit, new, with_amount}` don't return a `Result` anymore;
+- `{AliasOutput, BasicOutput, FoundryOutput, NftOutput}::{build_with_amount, build_with_minimum_storage_deposit}` don't return a `Result` anymore;
+- Lots of builder setters are now taking an `impl Into<Option<T>>` instead of a `T` parameter;
+
+### Removed
+
+- Remove `AddressGenerationOptions` in favor of `GenerateAddressOptions`, which now contains the `internal` flag;
 
 ## 0.2.0 - 2023-04-17